requirements.txt
requests==2.32.3
psycopg2==2.9.10
pymysql==1.1.1
sqlalchemy==2.0.36
pandas==2.2.3
google==3.0.0
mysql.connector==2.2.9
pytest==8.3.4 
pytest-mock==3.14.0

